rename misc_utils import to internal_utils
