+++ python/dgl/graphbolt/minibatch.py
@@ -7,7 +7,7 @@ import torch
 
 from .base import CSCFormatBase, etype_str_to_tuple, expand_indptr
 from .internal import get_attributes, get_nonproperty_attributes
-from .misc_utils import recursive_apply
+from .internal_utils import recursive_apply
 from .sampled_subgraph import SampledSubgraph
 
 __all__ = ["MiniBatch"]
